py_requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3